+++ core-bundle/src/Framework/ContaoFramework.php
@@ -399,7 +399,7 @@ class ContaoFramework implements ContaoFrameworkInterface
     {
         if (isset($GLOBALS['TL_HOOKS']['initializeSystem']) && is_array($GLOBALS['TL_HOOKS']['initializeSystem'])) {
             foreach ($GLOBALS['TL_HOOKS']['initializeSystem'] as $callback) {
-                System::importStatic($callback[0])->$callback[1]();
+                System::importStatic($callback[0])->{$callback[1]}();
             }
         }